updated-rest/pyproject.toml
[tool.poetry]
name = "karari"
version = "0.0.1"
description = ""
authors = ["ZeKoder <dev@zekoder.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
requests = "~2.28"
jsonschema = "^4.19.1"

# only when data section is there
pydantic = {extras = ["email"], version = "^1.10.4"}
sqlalchemy = "^1.2, !=1.2.9, < 1.4"
psycopg2-binary = "^2.9.3"
mongosql = "^2.0.15.post1"
dapr = "^1.8.3"
asyncpg = "^0.28.0"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"


[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"